requirements.txt
branca>=0.6.0
dtw_python>=1.3.0
folium>=0.14.0
geopandas>=0.14.0
hdbscan>=0.8.33
matplotlib>=3.8.0
numpy>=1.25.2
pandas>=2.1.3
pydantic>=2.5.1
rdp>=0.8
scikit_learn>=1.3.1
scipy>=1.11.4
Shapely>=2.0.2
tqdm>=4.66.1
